Inital restructure of app.py
