+++ girder_worker/app.py
@@ -1,12 +1,10 @@
-import json
 import sys
 
 import traceback as tb
 from distutils.version import LooseVersion
 
-import celery
-from celery import Celery, __version__
-from celery.result import AsyncResult
+from celery import Celery, __version__, current_app
+
 from celery.signals import (
     before_task_publish,
     task_failure,
@@ -15,186 +13,22 @@ from celery.signals import (
     task_revoked,
     task_success,
     worker_ready)
-from celery.task.control import inspect
+
 
 from girder_client import GirderClient
 
 import girder_worker
 from girder_worker import logger
-
-from girder_worker_utils import _walk_obj
-from girder_worker_utils.decorators import describe_function
+from girder_worker.context import get_context
+# Temporary
+from girder_worker.context.nongirder_context import MissingJobArguments
+from girder_worker.task import Task
+from girder_worker.utils import JobStatus, StateTransitionException, is_revoked
 
 import jsonpickle
 from kombu.serialization import register
-import requests
+
 import six
-
-from .utils import JobStatus, StateTransitionException
-
-
-class GirderAsyncResult(AsyncResult):
-    def __init__(self, *args, **kwargs):
-        self._job = None
-        super(GirderAsyncResult, self).__init__(*args, **kwargs)
-
-    @property
-    def job(self):
-        if self._job is None:
-            try:
-                # GirderAsyncResult() objects may be instantiated in
-                # either a girder REST request, or in some other
-                # context (e.g. from a running girder_worker instance
-                # if there is a chain).  If we are in a REST request
-                # we should have access to the girder package and can
-                # directly access the database If we are in a
-                # girder_worker context (or even in python console or
-                # a testing context) then we should get an ImportError
-                # and we can make a REST request to get the
-                # information we need.
-                from girder.utility.model_importer import ModelImporter
-                job_model = ModelImporter.model('job', 'jobs')
-
-                try:
-                    return job_model.findOne({'celeryTaskId': self.task_id})
-                except IndexError:
-                    return None
-
-            except ImportError:
-                # Make a rest request to get the job info
-                return None
-
-        return self._job
-
-
-class Task(celery.Task):
-    """Girder Worker Task object"""
-
-    _girder_job_title = '<unnamed job>'
-    _girder_job_type = 'celery'
-    _girder_job_public = False
-    _girder_job_handler = 'celery_handler'
-    _girder_job_other_fields = {}
-
-    # These keys will be removed from apply_async's kwargs or options and
-    # transfered into the headers of the message.
-    reserved_headers = [
-        'girder_client_token',
-        'girder_api_url',
-        'girder_result_hooks']
-
-    # These keys will be available in the 'properties' dictionary inside
-    # girder_before_task_publish() but will not be passed along in the message
-    reserved_options = [
-        'girder_user',
-        'girder_job_title',
-        'girder_job_type',
-        'girder_job_public',
-        'girder_job_handler',
-        'girder_job_other_fields']
-
-    def AsyncResult(self, task_id, **kwargs):
-        return GirderAsyncResult(task_id, backend=self.backend,
-                                 task_name=self.name, app=self.app, **kwargs)
-
-    def apply_async(self, args=None, kwargs=None, task_id=None, producer=None,
-                    link=None, link_error=None, shadow=None, **options):
-
-        # Pass girder related job information through to
-        # the signals by adding this information to options['headers']
-        # This sets defaults for reserved_options based on the class defaults,
-        # or values defined by the girder_job() dectorator
-        headers = {
-            'girder_job_title': self._girder_job_title,
-            'girder_job_type': self._girder_job_type,
-            'girder_job_public': self._girder_job_public,
-            'girder_job_handler': self._girder_job_handler,
-            'girder_job_other_fields': self._girder_job_other_fields,
-        }
-
-        # Certain keys may show up in either kwargs (e.g. via
-        # .delay(girder_token='foo') or in options (e.g.
-        # .apply_async(args=(), kwargs={}, girder_token='foo') For
-        # those special headers, pop them out of kwargs or options and
-        # put them in headers so they can be picked up by the
-        # before_task_publish signal.
-        for key in self.reserved_headers + self.reserved_options:
-            if kwargs is not None and key in kwargs:
-                headers[key] = kwargs.pop(key)
-            if key in options:
-                headers[key] = options.pop(key)
-
-        if 'headers' in options:
-            options['headers'].update(headers)
-        else:
-            options['headers'] = headers
-
-        return super(Task, self).apply_async(
-            args=args, kwargs=kwargs, task_id=task_id, producer=producer,
-            link=link, link_error=link_error, shadow=shadow, **options)
-
-    @property
-    def canceled(self):
-        """
-        A property to indicate if a task has been canceled.
-
-        :returns True is this task has been canceled, False otherwise.
-        """
-        return is_revoked(self)
-
-    def describe(self):
-        return describe_function(self.run)
-
-    def call_item_task(self, inputs, outputs={}):
-        return self.run.call_item_task(inputs, outputs)
-
-    def _maybe_transform_result(self, idx, result, **kwargs):
-        try:
-            grh = self.request.girder_result_hooks[idx]
-            if hasattr(grh, 'transform') and \
-               six.callable(grh.transform):
-                return grh.transform(result, **kwargs)
-            return result
-        except IndexError:
-            return result
-
-    def _maybe_transform_argument(self, arg, **kwargs):
-        if hasattr(arg, 'transform') and six.callable(arg.transform):
-            return arg.transform(**kwargs)
-        return arg
-
-    def _maybe_cleanup(self, arg, **kwargs):
-        if hasattr(arg, 'cleanup') and six.callable(arg.cleanup):
-            arg.cleanup(**kwargs)
-
-    def __call__(self, *args, **kwargs):
-        try:
-            _t_args = _walk_obj(args, self._maybe_transform_argument)
-            _t_kwargs = _walk_obj(kwargs, self._maybe_transform_argument)
-
-            results = super(Task, self).__call__(*_t_args, **_t_kwargs)
-
-            if hasattr(self.request, 'girder_result_hooks'):
-                if isinstance(results, tuple):
-                    results = tuple([self._maybe_transform_result(i, r)
-                                     for i, r in enumerate(results)])
-                else:
-                    results = self._maybe_transform_result(0, results)
-
-            return results
-        finally:
-            _walk_obj(args, self._maybe_cleanup)
-            _walk_obj(kwargs, self._maybe_cleanup)
-
-
-def _maybe_model_repr(obj):
-    if hasattr(obj, '_repr_model_') and six.callable(obj._repr_model_):
-        return obj._repr_model_()
-    return obj
-
-
-class MissingJobArguments(RuntimeError):
-    pass
 
 
 @before_task_publish.connect  # noqa: C901
@@ -203,99 +37,19 @@ def girder_before_task_publish(sender=None, body=None, exchange=None,
                                declare=None, retry_policy=None, **kwargs):
 
     try:
+        context = get_context()
         if 'jobInfoSpec' not in headers:
-            try:
-                # Note: If we can import these objects from the girder packages we
-                # assume our producer is in a girder REST request. This allows
-                # us to create the job model's directly. Otherwise there will be an
-                # ImportError and we can create the job via a REST request using
-                # the jobInfoSpec in headers.
-                from girder.utility.model_importer import ModelImporter
-                from girder.plugins.worker import utils
-                from girder.api.rest import getCurrentUser
-
-                job_model = ModelImporter.model('job', 'jobs')
-
-                user = headers.pop('girder_user', getCurrentUser())
-
-                # Sanitize any Transform objects
-                task_args = tuple(_walk_obj(body[0], _maybe_model_repr))
-                task_kwargs = _walk_obj(body[1], _maybe_model_repr)
-
-                job = job_model.createJob(
-                    **{'title': headers.pop('girder_job_title',
-                                            Task._girder_job_title),
-                       'type': headers.pop('girder_job_type',
-                                           Task._girder_job_type),
-                       'handler': headers.pop('girder_job_handler',
-                                              Task._girder_job_handler),
-                       'public': headers.pop('girder_job_public',
-                                             Task._girder_job_public),
-                       'user': user,
-                       'args': task_args,
-                       'kwargs': task_kwargs,
-                       'otherFields': dict(
-                           celeryTaskId=headers['id'],
-                           **headers.pop('girder_job_other_fields',
-                                         Task._girder_job_other_fields))})
-
-                headers['jobInfoSpec'] = utils.jobInfoSpec(job)
-
-            except ImportError:
-                # Current task will be the parent task in chain case
-                parent_task = app.current_task
-                try:
-                    if parent_task is None:
-                        raise MissingJobArguments('Parent task is None')
-                    if parent_task.request is None:
-                        raise MissingJobArguments("Parent task's request is None")
-                    if not hasattr(parent_task.request, 'girder_api_url'):
-                        raise MissingJobArguments(
-                            "Parent task's request does not contain girder_api_url")
-                    if not hasattr(parent_task.request, 'girder_client_token'):
-                        raise MissingJobArguments(
-                            "Parent task's request does not contain girder_client_token")
-                    if not hasattr(parent_task.request, 'id'):
-                        raise MissingJobArguments(
-                            "Parent task's request does not contain id")
-                    if 'id' not in headers:
-                        raise MissingJobArguments('id is not in headers')
-
-                    gc = GirderClient(apiUrl=parent_task.request.girder_api_url)
-                    gc.token = parent_task.request.girder_client_token
-
-                    task_args = tuple(_walk_obj(body[0], _maybe_model_repr))
-                    task_kwargs = _walk_obj(body[1], _maybe_model_repr)
-                    parameters = {
-                        'title': headers.pop('girder_job_title', Task._girder_job_title),
-                        'type': headers.pop('girder_job_type', Task._girder_job_type),
-                        'handler': headers.pop('girder_job_handler', Task._girder_job_handler),
-                        'public': headers.pop('girder_job_public', Task._girder_job_public),
-                        'args': json.dumps(task_args),
-                        'kwargs': task_kwargs,
-                        'otherFields': json.dumps(
-                            dict(celeryTaskId=headers['id'],
-                                 celeryParentTaskId=parent_task.request.id,
-                                 **headers.pop('girder_job_other_fields',
-                                               Task._girder_job_other_fields)))
-                    }
-
-                    try:
-                        response = gc.post('job', parameters=parameters, jsonResp=False)
-                        if response.ok:
-                            headers['jobInfoSpec'] = response.json().get('jobInfoSpec')
-                    except requests.exceptions.RequestException as e:
-                        logger.warn('Failed to post job: {}'.format(e))
-
-                except MissingJobArguments as e:
-                    logger.warn('Girder job not created: {}'.format(str(e)))
+            context.handle_jobInfoSpec(sender=sender, body=body, exchange=exchange,
+                                       routing_key=routing_key, headers=headers,
+                                       properties=properties, declare=declare,
+                                       retry_policy=retry_policy, **kwargs)
 
         if 'girder_api_url' not in headers:
             try:
                 from girder.plugins.worker import utils
                 headers['girder_api_url'] = utils.getWorkerApiUrl()
             except ImportError:
-                parent_task = app.current_task
+                parent_task = current_app.current_task
                 try:
                     if parent_task is None:
                         raise MissingJobArguments('Parent task is None')
@@ -320,7 +74,7 @@ def girder_before_task_publish(sender=None, body=None, exchange=None,
                     token = token_model.createToken(scope=scope, user=getCurrentUser())
                 headers['girder_client_token'] = token['_id']
             except ImportError:
-                parent_task = app.current_task
+                parent_task = current_app.current_task
                 try:
                     if parent_task is None:
                         raise MissingJobArguments('Parent task is None')
@@ -494,40 +248,6 @@ def gw_task_revoked(sender=None, request=None, **rest):
     except JobSpecNotFound:
         logger.warn(
             'No jobInfoSpec. Unable to move \'%s\' into CANCELED state.')
-
-
-# Access to the correct "Inspect" instance for this worker
-_inspector = None
-
-
-def _worker_inspector(task):
-    global _inspector
-    if _inspector is None:
-        _inspector = inspect([task.request.hostname])
-
-    return _inspector
-
-
-# Get this list of currently revoked tasks for this worker
-def _revoked_tasks(task):
-    _revoked = _worker_inspector(task).revoked()
-
-    if _revoked is None:
-        return []
-
-    return _revoked.get(task.request.hostname, [])
-
-
-def is_revoked(task):
-    """
-    Utility function to check is a task has been revoked.
-
-    :param task: The task.
-    :type task: celery.app.task.Task
-    :return True, if this task is in the revoked list for this worker, False
-            otherwise.
-    """
-    return task.request.id in _revoked_tasks(task)
 
 
 register('girder_io', jsonpickle.encode, jsonpickle.decode,
